# openapi.yaml
openapi: 3.0.4
info:
  title: HA SQLite Cluster API
  description: API for executing queries, managing databases, snapshots, and replications in HA. Access [litesql.github.io/ha/](https://litesql.github.io/ha/) to read the full documentation.
  version: 0.8.0
tags:
  - name: Main Database
    description: Operations on the main database.
  - name: All Databases
    description: Operations on all databases.
paths:
  /query:
    post:
      summary: Query the main database.      
      operationId: queryMainDatabase
      tags:
        - Main Database
      requestBody:
        description: Payload for the query request.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"              
      responses:
        '200':
          description: Result of the query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
  /undo/{txcount}:
    post:
      summary: Undo the last N transactions on the main database.
      operationId: undoMainDatabase
      tags:
        - Main Database
      parameters:
        - name: txcount
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Undo successful.
  /databases:
    get:
      summary: List all databases.
      operationId: listDatabases
      tags:
        - All Databases      
      responses:
        '200':
          description: A list of databases.
    post:
      summary: Create a database
      operationId: createDatabase
      tags:
        - All Databases
      requestBody:
        description: Payload for the create database.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDatabaseRequest"
      responses:
        '201':
          description: Created
  /databases/{id}:
    get:
      summary: Download a specific database.
      operationId: downloadDatabase
      tags:
        - All Databases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Database file.
    delete:
      summary: Delete a specific database.
      operationId: deleteDatabase
      tags:
        - All Databases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Database removed.
    post:
      summary: Query a specific database.
      operationId: queryDatabase
      tags:
        - All Databases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Payload for the query request.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        '200':
          description: Result of the query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
  /download:
    get:
      summary: Download the main database.
      operationId: downloadMainDatabase
      tags:
        - Main Database
      responses:
        '200':
          description: Main database file.
  /databases/{id}/snapshot:
    post:
      summary: Take a snapshot of a specific database.
      operationId: takeDatabaseSnapshot
      tags:
        - All Databases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Snapshot created.
    get:
      summary: Download a snapshot of a specific database.
      operationId: downloadDatabaseSnapshot
      tags:
        - All Databases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Snapshot file.
  /databases/{id}/undo/{txcount}:
    post:
      summary: Undo the last N transactions on a specific database.
      operationId: undoDatabase
      tags:
        - All Databases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: txcount
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Undo successful.
  /snapshot:
    post:
      summary: Take a snapshot of the main database.
      operationId: takeMainDatabaseSnapshot
      tags:
        - Main Database
      responses:
        '200':
          description: Snapshot created.
    get:
      summary: Download a snapshot of the main database.
      operationId: downloadMainDatabaseSnapshot
      tags:
        - Main Database
      responses:
        '200':
          description: Snapshot file.
  /databases/{id}/replications:
    get:
      summary: List replications for a specific database.
      operationId: listDatabaseReplications
      tags:
        - All Databases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of replications.
  /replications:
    get:
      summary: List all replication consumers for the main database.
      operationId: listMainReplications
      tags:
        - Main Database
      responses:
        '200':
          description: A list of replications.
  /databases/{id}/replications/{name}:
    get:
      summary: Get details of a specific replication for a database.
      operationId: getDatabaseReplication
      tags:
        - All Databases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Replication details.
    delete:
      summary: Delete a specific replication for a database.
      operationId: deleteDatabaseReplication
      tags:
        - All Databases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Replication deleted.
  /replications/{name}:
    get:
      summary: Get details of a specific replication.
      operationId: getReplication
      tags:
        - Main Database
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Replication details.
    delete:
      summary: Delete a specific replication.
      operationId: deleteReplication
      tags:
        - Main Database
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Replication deleted.
components:
  schemas:
    CreateDatabaseRequest:
      type: object
      properties:
        dsn:
          type: string
    QueryRequest:
      type: array
      items:
        type: object
        properties:
          sql:
            type: string
          params:
            type: object
            additionalProperties:
              oneOf:
                - type: string
                - type: integer
    QueryResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              columns:
                type: array
                items:
                  type: string
              rows:
                type: array
                items:
                  type: array
                  items:
                    type: string